New provider: DigitalOcean.
To use this provider, one additional environment variable is required:
* DO_PAT: DigitalOcean personal access token, required to update DNS.

The root domain name needs to be defined as well, of course.

Adding this provider required adding some additional packages to Godeps:
* github.com/digitalocean/godo
* github.com/tent/http-link-go
* golang.com/x/oauth2

One line in main.go to enable the new provider, and the provider itself.

I think I understand what's going on with Docker with this package now,
but I could be wrong so let me know what I need to change!