Set global TTL and fixed minor typo.
First, the typo:  it's DigitalOcean not Digital Ocean!

123456789 123456789 123456789 123456789 123456789 123456789 123456789 123456789
Next, the TTL issue.  Because records are set outside this provider,
I decided to remove the Provider's TTL attribute and instead set the
global TTL value in the config module.  This resolves the issue with
multiple TXT records for external-dns-adminproject in my domains.